# Replace storage. Move storage-related code from the chain.