--- a/core/storage/src/lib.rs
+++ b/core/storage/src/lib.rs
@@ -2,93 +2,37 @@ extern crate byteorder;
 extern crate elastic_array;
 #[cfg(test)]
 extern crate hex_literal;
-extern crate kvdb;
-extern crate kvdb_memorydb;
-extern crate kvdb_rocksdb;
 #[macro_use]
 extern crate log;
 extern crate primitives;
-extern crate serde;
-
-use std::collections::HashMap;
-use std::sync::Arc;
 
 pub use kvdb::{DBTransaction, DBValue, KeyValueDB};
-use kvdb_rocksdb::{Database, DatabaseConfig};
-use parking_lot::RwLock;
+use kvdb_rocksdb::{Database, DatabaseConfig, KeyValueDB};
 
-use primitives::traits::{Decode, Encode};
-
-pub mod trie;
-pub mod test_utils;
 pub mod storages;
+pub mod test_utils;
+pub mod trie;
 
-pub use trie::{DBChanges, Trie};
+use crate::storages::total_columns;
+use std::sync::Arc;
 pub use trie::update::{TrieUpdate, TrieUpdateIterator};
+pub use trie::{DBChanges, Trie};
+pub use storages::{BlockChainStorage, GenericStorage};
+pub use storages::beacon::BeaconChainStorage;
+pub use storages::shard::ShardChainStorage;
 
-pub const COL_STATE: Option<u32> = Some(0);
-pub const COL_EXTRA: Option<u32> = Some(1);
-pub const COL_BLOCKS: Option<u32> = Some(2);
-pub const COL_HEADERS: Option<u32> = Some(3);
-pub const COL_BLOCK_INDEX: Option<u32> = Some(4);
-pub const TOTAL_COLUMNS: Option<u32> = Some(5);
-
-pub type Storage = KeyValueDB;
-
-pub fn open_database(storage_path: &str) -> Database {
-    let storage_config = DatabaseConfig::with_columns(TOTAL_COLUMNS);
-    Database::open(&storage_config, storage_path).expect("Database wasn't open")
-}
-
-pub fn write_with_cache<T: Clone + Encode>(
-    storage: &Arc<Storage>,
-    col: Option<u32>,
-    cache: &RwLock<HashMap<Vec<u8>, T>>,
-    key: &[u8],
-    value: &T,
-) {
-    let data = Encode::encode(value).expect("Error serializing data");
-    cache.write().insert(key.to_vec(), value.clone());
-
-    let mut db_transaction = storage.transaction();
-    db_transaction.put(col, key, &data);
-    storage.write(db_transaction).expect("Database write failed");
-}
-
-pub fn extend_with_cache<T: Clone + Encode>(
-    storage: &Arc<Storage>,
-    col: Option<u32>,
-    cache: &RwLock<HashMap<Vec<u8>, T>>,
-    values: HashMap<Vec<u8>, T>,
-) {
-    let mut db_transaction = storage.transaction();
-    for (key, value) in values {
-        let data = Encode::encode(&value).expect("Error serializing data");
-        cache.write().insert(key.clone(), value.clone());
-        db_transaction.put(col, &key, &data);
+/// Initializes beacon and shard chain storages from the given path.
+pub fn create_storage(
+    storage_path: &str,
+    num_shards: u32,
+) -> (Arc<BeaconChainStorage>, Vec<Arc<ShardChainStorage>>) {
+    let db_config = DatabaseConfig::with_columns(Some(total_columns(num_shards)));
+    let db =
+        Arc::new(Database::open(&db_config, storage_path).expect("Failed to open the database"));
+    let beacon = Arc::new(BeaconChainStorage::new(db.clone()));
+    let mut shards = vec![];
+    for id in 0..num_shards {
+        shards.push(Arc::new(ShardChainStorage::new(db.clone(), id)));
     }
-    storage.write(db_transaction).expect("Database write failed");
-}
-
-pub fn read_with_cache<T: Clone + Decode>(
-    storage: &Arc<Storage>,
-    col: Option<u32>,
-    cache: &RwLock<HashMap<Vec<u8>, T>>,
-    key: &[u8],
-) -> Option<T> {
-    {
-        let read = cache.read();
-        if let Some(v) = read.get(key) {
-            return Some(v.clone());
-        }
-    }
-
-    match storage.get(col, key) {
-        Ok(Some(value)) => Decode::decode(value.as_ref()).ok().map(|value: T| {
-            let mut write = cache.write();
-            write.insert(key.to_vec(), value.clone());
-            value
-        }),
-        _ => None,
-    }
+    (beacon, shards)
 }