--- a/core/storage/src/lib.rs
+++ b/core/storage/src/lib.rs
@@ -14,7 +14,7 @@ pub mod storages;
 pub mod test_utils;
 pub mod trie;
 
-use crate::storages::total_columns;
+use crate::storages::NUM_COLS;
 use std::sync::Arc;
 pub use trie::update::{TrieUpdate, TrieUpdateIterator};
 pub use trie::{DBChanges, Trie};
@@ -28,7 +28,7 @@ pub fn create_storage(
     storage_path: &str,
     num_shards: u32,
 ) -> (Arc<RwLock<BeaconChainStorage>>, Vec<Arc<RwLock<ShardChainStorage>>>) {
-    let db_config = DatabaseConfig::with_columns(Some(total_columns(num_shards)));
+    let db_config = DatabaseConfig::with_columns(Some(NUM_COLS));
     let db =
         Arc::new(Database::open(&db_config, storage_path).expect("Failed to open the database"));
     let beacon = Arc::new(RwLock::new(BeaconChainStorage::new(db.clone())));