--- a/core/storage/src/storages/shard.rs
+++ b/core/storage/src/storages/shard.rs
@@ -10,7 +10,6 @@ use primitives::transaction::TransactionResult;
 use std::collections::HashMap;
 use std::io;
 use std::sync::Arc;
-use std::sync::RwLock;
 
 /// Shard chain
 pub struct ShardChainStorage {