… core/storage/Cargo.toml
… core/storage/benches/storage_bench.rs
… core/storage/benches/trie_bench.rs
… core/storage/src/storages/mod.rs
… core/storage/src/test_utils.rs
… core/storage/src/trie/mod.rs
… core/storage/src/trie/update.rs
… node/beacon/src/authority.rs
… node/beacon/src/beacon_chain.rs
… node/client/docs/client_storage.puml
… node/client/src/lib.rs
… node/client/src/test_utils.rs
… node/configs/src/authority.rs
… node/runtime/src/lib.rs
… node/runtime/src/test_utils.rs
… node/shard/src/lib.rs
… tests/test_rpc_cli.rs