* Architecture changes, see justification below:
    * Replaced old storage code with the new one;
    * Moved storage-specific code from chains;
    * Beacon and shard chain now share columns (had to merge the columns because of the bug in kvdb wrapper);
* Integration tests do not interfere anymore;
* Performance optimizations:
    * DevNet performance got visibly faster. No direct measurement, but previously the first transaction issued to DevNet would sometimes take 3-7 secs. Now it is instantaneous;
    * Storage code does not perform unnecessary cloning of the objects. Also the lookup of the existing object does not clone it (if it is in cache), unlike before;
    * Integration tests got 3-4x faster because they run in parallel;

Fixed #391 . Fixed #300. Fixed #179 .

### Architecture changes
https://www.dropbox.com/s/577lq3apprars8f/Screenshot%202019-02-18%2012.03.24.png
The proposed architecture achieves the following two major goals:
1. Separation of storage-specific logic and domain-specific logic;
2. Sequentialization of beacon chain together with beacon storage and shard chain together with shard storage;

Storage-specific logic includes:
* Cache;
* Code that integrates a specific db library. (kvdb wrapper has already proved to be buggy, so there is a good chance that we might decide to swap it for something later);
* Maintaining storage consistency. It is possible that an interrupted client might leave behind an inconsistent storage, e.g.: the blocks column has blocks for which we do not have corresponding headers in the headers column; or also more complex cases with column that represents indices might happen; The code that maintains storage consistency should be isolated;

Domain-specific logic includes:
* Fork-choice rule;
* Keeping track of finalized and pending blocks;
* etc.

There is a very high chance that beacon chain and each of the shard chains are going to be communicated to in parallel. Therefore it is very desirable for them to be sequentialized. However this sequentialization should also include the cache and the storage they are operating on. Therefore it is desirable to have beacon and shard -specific storages that can be sequentialized together with the corresponding chains.