@nearmax 2019-02-18
> Try rerun benchmarks

@evgenykuzyakov the storage benchmark bypasses most of the storage functionality because it uses the trie only. The real improvement comes in the cache implementation that the benchmark does not use. However, here is the comparison of old vs new results:
Old:
test trie_lookup ... bench:     449,039 ns/iter (+/- 35,359)
test trie_update ... bench:     356,865 ns/iter (+/- 63,417)

New:
test trie_lookup ... bench:     533,775 ns/iter (+/- 37,251)
test trie_update ... bench:     351,627 ns/iter (+/- 44,894)