@evgenykuzyakov 2019-02-18
@nearmax I was thinking about wasm benchmarks. But they also don't use storage as much. I probably worth it to create a storage heavy benchmark. Something in order of 250 or 500 reads.