@evgenykuzyakov 2019-02-18
Added wasm storage benchmarks in #584