@nearmax 2019-02-19
Added benchmark. Divide the following by 10.
```
test storage_save_block     ... bench:     229,673 ns/iter (+/- 80,900)
test storage_save_get_block ... bench:       2,005 ns/iter (+/- 304)
```